adding node v0.10 support via gnode
