+++ package.json
@@ -10,7 +10,8 @@
   },
   "dependencies": {
     "jade": "^1.3.0",
-    "co-fs": "^1.2.0"
+    "co-fs": "^1.2.0",
+    "gnode": "0.0.8"
   },
   "devDependencies": {
     "co-mocha": "0.0.4",
