+++ test/mocha.opts
@@ -1,0 +1,2 @@
+--require co-mocha
+--reporter spec
